Implement a dataset loading benchmark for Elasticsearch

+++ vdbbench/runner.py
@@ -95,14 +95,17 @@ def execute_runner(name: str, config: dict) -> dict:
         config_json_path = "/tmp/vdbbench/config.json"
         conn.put(io.BytesIO(config_json.encode()), config_json_path)
 
-        conn.run("sudo apt-get update")
-        conn.run("sudo apt-get install -y python3-pip python3-venv")
+        if conn.run("test -f /tmp/init_done", warn=True).failed:
+            conn.run("sudo apt-get update")
+            conn.run("sudo apt-get install -y python3-pip python3-venv")
+            conn.run("touch /tmp/init_done")
+
         conn.run("python3 -m venv /tmp/vdbbench/venv")
         conn.run(
             f". /tmp/vdbbench/venv/bin/activate && \
-              pip install -r /tmp/vdbbench/requirements.txt && \
-              cd /tmp/vdbbench && \
-              python -m vdbbench run-bench {name} {config_json_path}",
+            pip install -r /tmp/vdbbench/requirements.txt && \
+            cd /tmp/vdbbench && \
+            python -m vdbbench run-bench {name} {config_json_path}",
         )
 
         output_path = "/tmp/vdbbench/output.json"
